Create Hello World
Connect one end of a 330Ω resistor to PIN9 of the imp.
Connect the positive (long) lead of the LED to the other end of the resistor.
Connect the negative (short) lead of the LED to the GND pin of the imp (in this example, we've connected the lead through a wire).

Create a New Model
We assume you have already gone through the BlinkUp section of this guide. If you haven't, you should go do that now.

Log into the IDE
Click on New Devices to expand the drop down. You should see the device you've BlinkedUp.
Click on the device to open the Device Settings window.
If you want, you can give your device a name (*note - this name will be associated with the April board you plugged your imp into, not the imp itself*). We'll call our's april 1.
Click on the drop down box below Associated model and type the name for the new model you want to create. We'll call ours "Hello World."
Click Save Changes.
You should see the model you just created show up in the left hand nav bar. If you click on it, and then on your device, it will open the code editor.
The Code Editor
The code editor consists of three windows (agent, device, and logs):

Agent
The agent window is where you write your agent code. Agents are bits of Squirrel code that run in the Imp Cloud, and allow you to do things like offload heavy processing, make and processing HTTP requests, and store small amounts of data. You can learn more about agents in our next example.

Device
The device window is where you write your code that will run on the device itself. Since this example isn't going to use agents, this is where all of our code will go.

Logs
The logs window (at the bottom of the screen) displays various bits of information from the currently selected device. This includes system information (device online, offline, firmware updates, new code, etc), information about run time errors, and anything that was logged with server.log.